move attiny-85 to distinct module
