+++ mcu/attiny-hal/src/port.rs
@@ -1,0 +1,15 @@
+#[cfg(feature = "attiny85")]
+avr_hal_generic::impl_port_traditional! {
+    enum Ports {
+        PORTB: (crate::pac::PORTB, portb, pinb, ddrb),
+    }
+
+    pub struct Pins {
+        pb0: PB0 = (crate::pac::PORTB, PORTB, 0, portb, pinb, ddrb),
+        pb1: PB1 = (crate::pac::PORTB, PORTB, 1, portb, pinb, ddrb),
+        pb2: PB2 = (crate::pac::PORTB, PORTB, 2, portb, pinb, ddrb),
+        pb3: PB3 = (crate::pac::PORTB, PORTB, 3, portb, pinb, ddrb),
+        pb4: PB4 = (crate::pac::PORTB, PORTB, 4, portb, pinb, ddrb),
+        pb5: PB5 = (crate::pac::PORTB, PORTB, 5, portb, pinb, ddrb),
+    }
+}
